moved all the state and model functions closer to the ui they relate to
