--- src/actionForState.tsx
@@ -1,5 +1,0 @@
-import { renderApp } from "./renderApp";
-export const actionForState = state => stateHandler => (...args) => {
-    const newState = stateHandler(state, ...args);
-    renderApp(newState);
-};
