+++ src/components/castleRisk/updateState.ts
