+++ src/components/facebook/onFacebookLoginComplete.tsx
@@ -1,6 +1,6 @@
 import Cookies from "js-cookie";
-import { facebookUserSessionCookieKey } from "../facebookUserSessionCookieKey";
-import { inDevMode } from "../inDevMode";
+import { facebookUserSessionCookieKey } from "./facebookUserSessionCookieKey";
+import { inDevMode } from "../developers/inDevMode";
 
 export const onFacebookLoginComplete = (state, facebookUserSession) => {
     Cookies.set(facebookUserSessionCookieKey, facebookUserSession);
