+++ src/components/facebook/onFacebookLoginFailure.tsx
@@ -1,0 +1,2 @@
+import { identity } from "../../utils/identity";
+export const onFacebookLoginFailure = identity;
