+++ src/components/facebook/onLogoutRequested.tsx
@@ -1,5 +1,5 @@
 import { deleteSessionCookie } from "./deleteSessionCookie";
-import { inDevMode } from "./inDevMode";
+import { inDevMode } from "../developers/inDevMode";
 
 export const onLogoutRequested = (state) => {
     const { facebookUserSession, ...remainingState } = state;
