+++ src/components/nav/navTo.tsx
@@ -1,4 +1,4 @@
-import { history } from "../history";
+import { history } from "./history";
 
 export const navTo = (state, path: string) => {
     const url = state.url.setPath(path);
