+++ src/components/organizations/addNewOrganization.tsx
@@ -1,4 +1,5 @@
-import { navTo } from "../../models/navTo";
+import { navTo } from "../nav/navTo";
+
 export const addNewOrganization = (state, creatorEmail) => {
     const { organizations } = state;
     const href = `/organizations/${organizations.length + 1}`;
