+++ src/components/state/makeStatefulCallback.tsx
@@ -1,0 +1,6 @@
+import { renderApp } from "./renderApp";
+
+export const makeStatefulCallback = (state, stateUpdatingCallback) => (...args) => {
+    const newState = stateUpdatingCallback(state, ...args);
+    renderApp(newState);
+};
