--- src/heartpoints-cli.ts
@@ -1,4 +1,0 @@
-import { allStates } from "./models/economy-model";
-import { jsonString } from "./utils/jsonString";
-
-console.log(jsonString(allStates));
