--- src/navigateToSimpleModel.tsx
@@ -1,1 +1,0 @@
-export const navigateToSimpleModel = (state) => ({ ...state, showSimpleModel: true });
