--- src/newSiteProps.tsx
@@ -1,8 +1,0 @@
-import { actionForState } from "./actionForState";
-import { mapProperties } from "./utils/list";
-import { actionables } from "./actionables";
-
-export const newSiteProps = state => ({
-    ...state,
-    ...mapProperties(actionables, actionForState(state)),
-});
