--- src/onFacebookLoginFailure.tsx
@@ -1,2 +1,0 @@
-import { identity } from "./utils/identity";
-export const onFacebookLoginFailure = identity;
