+++ src/utils/abstractAlgebra.ts
