+++ src/utils/range.ts
@@ -1,0 +1,1 @@
+export const range = (n: number): number[] => n == 0 ? [0] : [n, ...range(n - 1)];
